+++ bindings/tests/bindings/reverse.py
@@ -58,7 +58,7 @@ async def swap(to_chain: kaleidoswap_sdk.Chain):
             fee=kaleidoswap_sdk.Fee.ABSOLUTE(200),
             output_address=claim_address,
             chain_client=chain_client,
-            kaleidoswap_sdk=boltz_api,
+            boltz_api=boltz_api,
         ),
     )